Complete fix 4

+++ ahrameev.vladislav/P3/matrix.h
@@ -6,12 +6,9 @@
 namespace ahrameev
 {
     bool checkArguments(int argc, char* argv[]);
-
-    bool readMatrixSize(std::ifstream& input, int* rows, int* cols);
-
-    void processSpiralDecrease(const int* src, int rows, int cols, std::ofstream& output);
-
-    void processLowerTriangle(const int* matrix, int rows, int cols, std::ofstream& output);
+    bool readMatrixSize(std::ifstream& input, size_t* rows, size_t* cols);
+    void processSpiralDecrease(const int* src, size_t rows, size_t cols, std::ofstream& output);
+    void processLowerTriangle(const int* matrix, size_t rows, size_t cols, std::ofstream& output);
 }
 
 #endif
